+++ src/enginehandleproducer.rs
@@ -1,7 +1,6 @@
-use hirofa_utils::single_threaded_event_queue::SingleThreadedEventQueue;
+use hirofa_utils::eventloop::EventLoop;
 use mozjs::rust::{JSEngine, JSEngineHandle};
 use std::cell::RefCell;
-use std::sync::Arc;
 
 // this class has a single taskmanager which has a single thread which initializes and destroys the JSEngine
 // one might argue that you should just do this from your main thread but that does not work with tests
@@ -11,11 +10,11 @@ thread_local! {
 }
 
 lazy_static! {
-    static ref EVENTQUEUE: Arc<SingleThreadedEventQueue> = SingleThreadedEventQueue::new();
+    static ref EVENTLOOP: EventLoop = EventLoop::new();
 }
 
 pub fn produce() -> JSEngineHandle {
-    EVENTQUEUE.clone().exe_task(|| {
+    EVENTLOOP.exe(|| {
         ENGINE.with(|engine_rc| {
             let engine: &JSEngine = &*engine_rc.borrow();
             engine.handle()